add feed flow and rename view directories to view_model
