+++ frontend/lib/ui/feed/view_model/feed.dart
@@ -1,0 +1,1 @@
+/* Retrieve feed from the feed repository */
