+++ frontend/lib/ui/feed/widgets/feed_screen.dart
@@ -1,0 +1,1 @@
+/* Display feed from the feed view model */
